Fix to work with puppetlabs-concat 2.0.0+.
The concat module changed behaviour in 2.0.0 such that concats without
any fragments generate an error:

"The fragments directory is empty, cowardly refusing to make empty config files"

One fix would have been to enable the force option. But this wouldn't
have been backwards compatible; older versions of the module would
generate a deprecation warning (force was deprecated prior to 2.0.0
and then undeprecated in 2.0.0).

So this fix simply adds a gratuitious header to the files which fixes
the problem on 2.0.0+ and doesn't cause an issue with older versions.